#913: ToDoListGroup 공통 엔티티로 분리
