+++ ToDo-Garden/ToDo-Garden/EditToDoScene/Sources/EditToDoSceneAPI/EditToDoSceneBuildable.swift
@@ -9,7 +9,6 @@ import Foundation
 
 /// 런타임에 전달받을 의존성을 선언한 구조체입니다.
 public protocol EditToDoScenePayloadable {
-  var toDoId: UUID { get }
 }
 
 @MainActor
