+++ ToDo-Garden/ToDo-Garden/HomeScene/Sources/HomeSceneEntity/HomeSceneModels.swift
@@ -22,9 +22,9 @@ public enum HomeScene {
     
     public struct Response: Codable, Sendable {
       public let date: String
-      public let list: [TodoListGroup]
-      
-      public init(date: String, list: [TodoListGroup]) {
+      public let list: [SharedEntity.TodoListGroup]
+
+      public init(date: String, list: [SharedEntity.TodoListGroup]) {
         self.date = date
         self.list = list
       }
@@ -38,54 +38,6 @@ public enum HomeScene {
       public init(data: [TodoBatchItem]) {
         self.data = data
       }
-    }
-  }
-}
-
-// MARK: DTO
-extension HomeScene {
-  public final class TodoListGroup: Codable, @unchecked Sendable {
-    public let localId: String
-    public let name: String
-    public let color: String
-    public var todoList: [TodoListItem]?
-    public let progressRate: Double
-    
-    public init(localId: String, name: String, color: String, todoList: [TodoListItem]?, progressRate: Double) {
-      self.localId = localId
-      self.name = name
-      self.color = color
-      self.todoList = todoList
-      self.progressRate = progressRate
-    }
-  }
-  
-  public final class TodoListItem: Codable, @unchecked Sendable {
-    public var name: String
-    public let endDay: String?
-    public var isDone: Bool
-    public let localID: String
-    public let startDay: String?
-    public let alarmTime: Int?
-    public let isAlarmOn: Bool
-    public let isOnlyToday: Bool
-    public let repeatToDoId: String?
-    
-    enum CodingKeys: String, CodingKey {
-      case name, endDay, isDone, startDay, alarmTime, isAlarmOn, isOnlyToday, repeatToDoId
-      case localID = "local_id"
-    }
-    
-    public init(name: String, endDay: String?, isDone: Bool, localID: String, startDay: String?, alarmTime: Int?, isAlarmOn: Bool, isOnlyToday: Bool, repeatToDoId: String?) {
-      self.name = name
-      self.endDay = endDay
-      self.isDone = isDone
-      self.localID = localID
-      self.startDay = startDay
-      self.alarmTime = alarmTime
-      self.isAlarmOn = isAlarmOn
-      self.isOnlyToday = isOnlyToday
-      self.repeatToDoId = repeatToDoId
     }
   }
 }
